% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{checkStopVerbose}
\alias{checkStopVerbose}
\title{checkStopVerbose}
\usage{
checkStopVerbose(muold, mu, sigmaold, sigma, err, it, maxit, verbose)
}
\description{
check whether or not to stop (i.e. convergence)
}
\keyword{internal}
