% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{checkStopVerboseMulti}
\alias{checkStopVerboseMulti}
\title{checkStopVerboseMulti}
\usage{
checkStopVerboseMulti(muold, mu, sigmaold, sigma, err, it, maxit, verbose)
}
\description{
check whether or not to stop (i.e. convergence) (multivariate)
}
\keyword{internal}
