% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{initialIndicesMulti}
\alias{initialIndicesMulti}
\title{initialIndicesMulti}
\usage{
initialIndicesMulti(
  y,
  mu,
  sigma,
  k,
  dampFactor = NULL,
  forceDetectDamp = FALSE
)
}
\description{
initialize indices for multivariate scenario
}
\keyword{internal}
