% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfishHmrf.hmrfem.R
\name{smfishHmrf.hmrfem}
\alias{smfishHmrf.hmrfem}
\title{smfishHmrf.hmrfem}
\usage{
smfishHmrf.hmrfem(
  y,
  neighbors,
  numnei,
  blocks,
  beta = 0.5,
  mu,
  sigma,
  err = 1e-04,
  maxit = 20,
  verbose
)
}
\arguments{
\item{y}{gene expression matrix}

\item{neighbors}{adjacency matrix between cells}

\item{numnei}{an array of number of neighbors per cell}

\item{blocks}{a list of cell colors for deciding the order of cell update}

\item{beta}{the beta to try (smoothness parameter)}

\item{mu}{an array of cluster mean}

\item{sigma}{an array of cluster standard deviation}

\item{err}{the error that is allowed between successive iterations}

\item{maxit}{maximum number of iterations}

\item{verbose}{TRUE or FALSE}
}
\value{
A list of prob, new mu, and new sigma after iterations finish
}
\description{
HMRF for single variate normal distribution
}
\keyword{internal}
