% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smfishHmrf.hmrfem.R
\name{smfishHmrf.read.expression}
\alias{smfishHmrf.read.expression}
\title{smfishHmrf.read.expression}
\usage{
smfishHmrf.read.expression(expr_file)
}
\arguments{
\item{expr_file}{expression matrix file. The expression file should be a space-separated file. The rows are genes. The columns are cells. There is no header row. The first column is a gene index (ranges from 1 to the number of genes). Note the first column is not gene name.}
}
\value{
A matrix with gene expression matrix
}
\description{
Reads expression
}
\examples{
expr_file = system.file("extdata", "ftest.expression.txt", package="smfishHmrf")
y<-smfishHmrf.read.expression(expr_file)
}
